refactor: move csrf function to subfolder csrf
