+++ app/routes/shared/components/VorabcheckPage.tsx
@@ -7,7 +7,7 @@ import { ProgressBar } from "~/components/form/ProgressBar";
 import PageContent from "~/components/PageContent";
 import { StrapiFormComponents } from "~/services/cms/components/StrapiFormComponents";
 import { splatFromParams } from "~/services/params";
-import { CSRFKey } from "~/services/security/csrfKey";
+import { CSRFKey } from "~/services/security/csrf/csrfKey";
 import { validatorForFieldnames } from "~/services/validation/buildStepValidator";
 import type { loader } from "../vorabcheck.server";
