+++ app/routes/shared/formular.server.ts
@@ -26,7 +26,7 @@ import { navItemsFromStepStates } from "~/services/flowNavigation.server";
 import { logError } from "~/services/logging";
 import { stepMeta } from "~/services/meta/formStepMeta";
 import { parentFromParams } from "~/services/params";
-import { validatedSession } from "~/services/security/csrf.server";
+import { validatedSession } from "~/services/security/csrf/validatedSession.server";
 import {
   getSessionData,
   getSessionManager,
