+++ app/services/security/csrf/getCSRFFromSession.server.ts
@@ -1,0 +1,5 @@
+import type { Session } from "@remix-run/node";
+import { CSRFKey } from "./csrfKey";
+
+export const getCSRFFromSession = (session: Session) =>
+  session.get(CSRFKey) as string[] | null;
