+++ app/services/session.server/updateSessionInHeader.ts
@@ -1,6 +1,6 @@
 import type { FlowId } from "~/flows/flowIds";
 import { lastStepKey } from "~/services/flow/constants";
-import { createSessionWithCsrf } from "~/services/security/csrf.server";
+import { createSessionWithCsrf } from "~/services/security/csrf/createSessionWithCsrf.server";
 import {
   type CookieHeader,
   getSessionManager,
